Fixed contact edit/delete/change primary --again

+++ app/Http/Controllers/AccountController.php
@@ -214,9 +214,9 @@ class AccountController extends Controller {
                 $emailAddressRepo->Delete($emsToDelete);
         }
 
-        foreach($contactsToDelete as $delete_id) {
-            $contactRepo->Delete($delete_id);
-        }
+        if ($contactsToDelete !== null)
+            foreach($contactsToDelete as $delete_id)
+                $contactRepo->Delete($delete_id);
         //END contacts
 
         //BEGIN delivery address
